Fixing some bugs working out the total hours.

+++ lib/burndown/milestone_event.rb
@@ -4,7 +4,7 @@ module Burndown
     
     property :id,               Serial
     property :open_tickets,     Text,   :default => "" # comma-separated ids: 1,2,3
-    property :hours_left,       Integer, :default => 0.0
+    property :hours_left,       Float, :default => 0.0
     property :created_on,       Date
     
     belongs_to :milestone
